Tests/LibWeb/Text/expected/DOM/CDATASection-cloneNode.txt

Cloned CDATASection node data: PASS
